Temporarily disable errors in project structure dialog
For a gradle project, the project structure dialog could show bogus
errors. Not only are the errors incorrect, the quick fixes for them
also don't work.

Unfortunately, right now there is no API to enable/disable this
error checking per project. This CL adds a VM option that completely
disables this error checking in project structure. It should be
reverted as soon as possible.

Tracking bug to revert this CL is here:
https://code.google.com/p/android/issues/detail?id=63310

Change-Id: I670132c5464c847a49243cb671e214c1711b3471