Updater: add unit tests.
This adds 2 kinds of tests to the updater:
- a typical unit test that mocks the UI & the resolved jar.
- a "black box" large test that builds an updater with data
  and runs it. This is done from an external gradle task.

Invoking the Runner from unit tests requires to
change it to not always create a Swing UI but
pass a MockUpdaterUI object instead and also
to inject an interface to resolve the expected
updater.jar to somethig more suitable for testing.

Also adds a larger UI test to the updater, using it
as a "black box" by creating a patch between 2 mock
data folders and then applying the patch.

The UI test must be invoked using ./gradlew testUI
and since it will use some UI should not be invoked
on a CIS or test server.

SDK Bug: b.android.com/63927

Change-Id: I2630aa9cb838c2b0d83261fd4d8643554a5a345f