Gracefully handle build.gradle files in modules without a configured JDK
The search scope set up for gradle files needs to have a valid project
(because it is used to look up JavaTypes during type attribution). If we're
setting up a build script file in a module, make sure we include a search
scope with a valid project such that we don't hit an NPE later.

Change-Id: Ic644596052ded1638ef15f6e5e1e42dc9f00315f