Include ui-designer-core inside Android plugin.
A recent change (18aad9f39f) moved all the sources of
android-designer plugin into the android plugin. The
android-designer plugin wasn't removed since removing
it might introduce update issues.

This CL updates the build scripts so that the generated
IDE continues to work after the above update. Specifically,
the android-designer plugin included all the sources
and resources of the ui-designer-core plugin. These should
now be included in the android plugin. In order to do this,
this CL unwraps the utility method layoutPlugin("android")
into its constituent functions so that sources from
multiple modules (android and ui-designer-core) can be
included inside the android plugin.

This CL does not touch the layout of the android designer
plugin. We need to figure out if that plugin should continue
to include the designer core plugin.

Change-Id: I168556879fd710093f189747e3ed8d55a870ce17