Layout editor: Allow Mac ctrl-click on multi-selection
On the Mac, where it's common to have a single button on a mouse,
right clicks are handled by ctrl+click. However, the layout editor
would also treat the click as a selection event, meaning that if you
had selected multiple components, and tried to right click on one of
them to open a menu to operate on all of them (for example, to invoke
Refactor > Extract Include in the Android layout editor in order to
operate on the selected children), that click would result in the
selection getting replaced with just the component under the mouse
before the action was invoked.

Change-Id: I4d983bec5a120676248be6878f817290ef1d607e