Don't enable debugging on JVM running tests
While running tests on the build server, we often see the failure
message about unable to bind to a port since it is already in use.
This CL is an attempt to see if not passing the -Xdebug and -Xjdwp
options to the VM that is launched for running tests will help
avoid the issue.

Change-Id: I54bfb922d1de11c6f8689886342de29bf745f04f