Don't enable debugging during build
In a previous CL, the -Xdebug option was disabled while running
tests in order to prevent the tests from failing to run if multiple
builds run simultaneously and accessed the same port. However it
was still being used during the initial build part. This CL removes
it globally during all builds.

Change-Id: I982d827f6d6b299f63045406b41b39e28e65dc0b