72680: Handle Studio settings upgrades more automatically
When you run AndroidStudioBeta or AndroidStudio for the first time, make the
import helper aware of our previous settings directory names (AndroidStudioPreview,
AndroidStudioBeta) such that it can default to suggesting an import location which
migrates the settings. Prior to this fix (which unfortunately meant this happened in
beta) users would simply be pointed to the /Applications folder on Mac (probably
something similar on other platforms) and if they had already replaced their preview
install with beta, it wasn't clear what to do.

Change-Id: If004f2b652df51f11f47622ea991a6dd01885af5