Updater: Fixes revert leaving a broken build behind
When the updater fails or it is canceled it reverts all the actions
up to that point. When it fails in windows because a file is opened
by another process and cannot be deleted then the following happens:

  1. The updater fails to delete the file and starts the revert
     process.
  2. The revert process reverts the operations in REVERSE order,
     which means that the file that failed is the first one to be
     reverted.
  3. Reverting a delete operation does first a DELETE on the file
     before restoring it.

The delete in step 3 is needed in case another file was created in
place of this one at a later stage. However, in this case the file
cannot be deleted because its locked! An exception is thrown and
NO REVERT operations are applied.

This change simply adds a similar check that the other actions have,
that is if the file is already there and it's checksum is the same
one, then the revert is not needed.

This fixes the multiple issues with corrupted installations that we
have seen on Windows.

Change-Id: Id81f61294dae93a6152055a33c925562155c59fa