Updater: Adds check for process lock on files
The updater checks for read/write access to the files and
also performs a FileLock check on them. However, on Windows,
this is not enough to see if a file can be deleted. If
another process holds a lock to a file, neither mechanism
will detect it. This change uses a Win API via JNA to detect,
and offer the KILL_PROCESS solution, to the user.

Change-Id: I0823f0a70d568a9a2a0453e6ec96af9eaf76550c