Updater: Fixes delete action for binary patches.
Delete actions were using "zip" checksums even for binary patches.
This caused the Updater to flag the file to be deleted as modified
as the created checksum was for the whole file, but the calculated
on application was the "zip" one.

Note that DeleteAction was pretty much renamed to BaseDeleteAction,
but the diff might not pick this up correctly, as a new DeleteAction
was created.

Change-Id: I2b9e337556ec9f7d10285391e41bc55f1b0b930a