Updater: Fixes issue when reverting a directory delete.
Reverting a delete operation for a directory would cause the isModified
function to be called on it, and it's not intended to be run on
directories.

Adds tests for the case when a file is locked by another process and it used
to trigger a corrupted installation path as the revert was throwing
an exception.

Also updates failing tests to cope with the new KILL_PROCESS.

Change-Id: Iec78bc53d86a9826f47485a2380c5d33efa2c7a8
(cherry picked from commit 9229e54299821191b16766bd07b59cc6cb3f12fa)