Updater: Moves properties file to patch file.
When a jar for the updater is created, one file that belongs to the
patch is actually added to the jar instead of the patch file itself.
This is a properties file and currently only stores the versions
patching "from" and "to". However in future CLs this will contain
more metadata on the patch itself. For example if it needs to be
"strict" (ie. don't allow to proceed).

By moving the properties file to the patch.zip file we can separate
the code from the patch data.

Change-Id: I03548d60f2cc5075e14ed7334513dd4c1f21c905