Updater: Support for moving files, and multiple patch sources
This change also generalized update actions to be able to take the
data from a different source location. For this to happen update
actions now read from the backup folder to make sure that if there
were other actions applied to the source file, we still have the
original copy. This effectively implements an update-on-move
operation. For a mac example where we moved everything into the
contents folder this brings down patch size from 170MB to 18MB,
which is almost the size it would have had if nothing had moved.

This change also adds support for patching from multiple sources into
a single destination. This allows patching from different
installations of the same version, into a common one. This allows
us to unify different installations into one where its binary
signature can be verified.

This change achieves this by creating N patches from N different
variations of the source folder to the single target folder. The
N patches are then merged as follows:
  - All the actions that affect the same target path are merged.
  - If the actions to be merged are equal then only one remains.
  - If they are different a MultiAction containing them is created.

MultiActions work as follows:
  When the patch is created they create zip entries for all the
  variants they possess. (Alternative binary diff data for each
  variant are stored in the same patch file but under the
  directory /.variantN).

  When a patch is applied, first the current installation is
  validated. MultiActions use this time to select which of its
  actions needs to be applied. All its actions are validated and
  the one that doesn't return a validation error is chosen, and
  later applied.

Note that patching from multiple input sources will increase the
patch size, but as an example patching from the 8 different
variations of the latest beta installation to the current build
gives us a 20MB patch.

Change-Id: Id09f1c04021d848e4a1a9bdfc80261372ddfda5b