Updater: Critical files are always included in the patch.
The critical flag was previously unused. Now if a file is flagged as
critical it will always be detected as changed and will be included
in the patch. Moreover the whole content will be added so it
can be replaced at destination.

Change-Id: Ib01d1de3689ed6eed9ec10be2f96964ddbb0dc87
(cherry picked from commit 7bcc77577a970922bffa8fdaa3e26bf47118cccd)