Updater: Fix for normalization of jar files.
It turns out that the binary result of writing a ZIP file is
platform dependent. In particular the generic flag for
"Language encoding flag (EFS)." (bit 11 of the local header) changes
between mac and linux. Then a patch generated in linux cannot be
applied on Mac. In Java 7 there is a way of forcing this flag
when creating a Zip file by specifying the Charset on creation,
however that doesn't exist in Java 6.

This change removes completely the dependency from a zip file
and actually normalizes the data to a stream of one file after
the other. This for some reason suits better the idea.jar file
and the patch goes down from 98MB to 87MB.

Change-Id: I5139be567877f8efa291a6e29e3fee096e9973c2