Use custom color for files in the test scope
IntelliJ provides a setting (Settings | File Colors) that allows
customization of colors for files in different scopes. This setting
defaults to including a custom color for files in the test scope,
but that default is only available in certain IntelliJ products.
This CL enables that setting for Android Studio.

Part of the fix for https://code.google.com/p/android/issues/detail?id=78008

Change-Id: I8ed15c046e5c976a95af278b9487c3e4d1a670a4