Launcher: Fix for vmoptions lookup.
The way studio/idea finds the vmoptions file is inconsistent
between the batch file and the .exe launchers.

The batchfile will look first at the [PRODUCT]_VM_OPTIONS_FILE
environtment variable and if its not there it would then use
[product]%bits%.exe.vmoptions

whereas the .exe launcher will look frist at

(for studio) .AndroidStudioPreview/studio.vmoptions

This is CL 1/2 that changes the batchfile too also look at the
platform selector directory, so for android studio the search is:

ANDROID_STUDIO_VM_OPTIONS_FILE
%USERPROFILE%/.AndroidStudioBeta/studio%BITS%.exe.vmoptions
%BIN_DIR%/studio%BITS%.exe.vmoptions

CL 2/2 changes the native launcher to look in this order:

%USERPROFILE%/.AndroidStudioBeta/studio%BITS%.exe.vmoptions
%BIN_DIR%/studio%BITS%.exe.vmoptions

Which currently is looking for the user specific file at:
%USERPROFILE%/.AndroidStudioPreview/studio.vmoptions.

This way there is a unified way of telling our users where
to put the vmoptions file.

Change-Id: Ieea292b0d2a295b5ba06179fe212541c3f0c4356