Updater: Do not move a directory into a file.
We were incorrectly taking into account directories as 'sources'
for file moves. If such a directory was chosen, then it would
fail trying to read with this. This only occured if a file
with name "X" was created and at the same time a directory with name
"X/" was removed.

Change-Id: I35b427de3888c865d1695bf51cb9a1573fd98aaa