Load dlls from embedded jna library
JNA includes a platform specific shared library. According to the JNA
documentation, this library is loaded from:
 - directories specified in jna.boot.library.path
 - system library paths
 - or by extracting the stub library from from the JNA jar file
in that sequence. This setting jna.nosys=true disables loading from
the system and will force the last option. This CL enables that
setting for Studio and for the restarter launched by Studio.
This is an attempt to see if the user reports that specify the
following error:

 java.lang.UnsatisfiedLinkError: Can't obtain static newInstance \
 method for class com.sun.jna.Structure

are resolved by not picking some random dll from the system library
path.

See https://code.google.com/p/android/issues/detail?id=67091

Change-Id: Ib88142b50ba5dd6865dc8d30811df0e20ec114fd