The ide displays an error for valid call to getClass()
https://code.google.com/p/android/issues/detail?id=70135

This is an infrastructure change required for the actual fix.
It's also reflected at JetBrains tracker:
https://youtrack.jetbrains.com/issue/IDEA-132360

The problem is that either *.class or *.java files are in resolve
scope in android environment sometimes. The android plugin
already tries to solve that
(see AndroidSdkResolveScopeProvider.MyJdkScope.compare())
but sometimes the problem occurs (e.g. the mentioned case
is when current class extends a class from support library and
it's ancestor roots to an Object class from *.class file configured
at the android sdk. I.e. that Object from *.class comes into resolve
scope when we go up the class hierarchy. An Object class from
*.java file is already in the scope though, hence, we have a conflict)

Change-Id: Ia7149cd9ecdf81c1c10c4e550c350a8ab292dd18