Updater: Makes Mac patches start at the root folder
On mac patches were forced to start at the Contents fodler due to a
change on the application that made it return Contents as the home
folder. This however had uninteneded side effects:

 * Files outside the Contents folder can nolonger be patched. We
   currently don't have any, but we wouldn't be able to create one.
   We also cannot see any files outside Contents and won't be
   able to prompt for deletion.

 * There was a line that changed the timestamp of the .app bundle.
   This was done so the Info.pList cache was updated. However after
   the Contents move, the directory whose timestamp is updated is
   Contents and not the .app one.

This CL also fixes a small bug when detecting a move on a critical
file.

Note that the list of files for arguments like "critical" or
"warnings" is still relative to the <old_folder> directory.

Change-Id: Id9f383da7825470508b87dca763ceaa1f5ecea60