idea.properties: Load all and override properties
Currently if a user provides an idea.properties file, it completely
overrides the one in idea. This means that when that happens it is
no longer possible to update the original one and have any effect.

This change makes idea load all properties files in preference
order, so the user given properties have more priority than the
application ones.

Change-Id: Iaba11357f2dbe26b03518dce11e54ebf1b3f4313