Generate searchableOptions.xml at build time
There is a static, obsolete search index checked into the
repository (resources-en/src/search/searchableOptions.xml).

However, the actual metadata for search should be derived at build time,
since it should (1) only include the specific set of options available
from plugins included in Android Studio, (2) use the correct UI labels
(such as product name) configured at build time, and (3) pick up any
new options added in the future.

There was some support for this in the build scripts, but it was not
enabled in the community edition; however, the Python build scripts included
the rest of the setup. This CL does the same thing for the Studio build.

It also disables some additional plugins and jars not needed in Studio.

Change-Id: I8abdcae43812c78b209188dd317a75e10b0a679e