Add a hyper link label to the Usage Statistics dialog.
This CL adds a hyperlink label to the usage statistics dialog. This
allows extensions to provide hyper links, such as a link to the company's
privacy policy.

We cannot fold this into the existing label since JLabel's don't support
links, and we cannot fold the existing JLabel's into this hyperlink since
the HyperLinkText component doesn't support breaking of lines and hence is
unsuitable for a larger body of text.

Change-Id: I205ebb53b19a0c17d8e213b7ef440128cab60f07

Conflicts:
	platform/platform-impl/src/com/intellij/internal/statistic/configurable/StatisticsConfigurationComponent.java