Rearrange attributes in XML files by default
In ADT, we rearranged XML attributes automatically (e.g. placing
layout_width before layout_height even though alphabetically it comes
after.)

We couldn't do that in Android Studio because the IntelliJ formatter
had a global flag for whether a reformat should also rearrange
elements, so if we enabled it, it would also cause reformatting in
.java files to rearrange methods and fields -- something most users
don't want.

However, as of IntelliJ 14.1 the formatter tracks the rearrange state
per language. Therefore, we can now change the default for XML to
true, which this CL does.

Change-Id: Ie586188043a19aeb4166051f1580dbeabc928039