Minor tweak required to build on Windows.
Currently, a command-line build will fail to delete the out/ folder, because
it is already "in use" by Ant. The project's build log is written to
out\tmp\system\build-log\build.log, the JPS cache is initialized under
out\.jps-build-data folder, and then the script attempts to delete out. This
CL simply reorders the clean step.

Change-Id: If88c3cfd4cc9cf8f67ec26bdd1c7898bbb9b5eb1