[UiTests] Fix Windows command-line invocation (classpath too long).
This change mirrors the upstream fix for running Intellij's unit tests
on Windows, by writing the classpath to a temporary .classpath file. A
simple launcher (with no dependencies on community-main) is then used
to read in the contents of the classpath and bootstrap the actual
tests.

- gui_tests.gant now has a similar setup to common_tests.gant
- BootstrapUITests is similar to BootstrapTests, but note our tests
use custom runners injected via @RunWith, so we need JUnit4TestAdapter.

Change-Id: I0888cb065c1f8c399a2d767554704f71b88243f1