171122: Copy studio64.vmoptions during AndroidStudio update
It turns out that the "import settings" wizard only copies
data from the old config/ folder to the new one. But there
are a couple of optional files *outside* the config folder:
idea.properties, and the .vmoptions files. These were not
copied, as recorded on 171122. (This is not a problem on
OSX which uses the OS folders ~/Library/Preferences/<App>
both as the config folder and as the folder which hosts these
settings files.)

Also, the code was checking for the age of the config file
and picking the "newest" one. However, when files are copied
they (at least on Linux, where I observed this) keep the
same modification time as the source. So if settings had
been imported from 1.2-preview to 1.2 final, they had the
same modification timestamp, and since the comparison only
look for >, not >=, it ended up using the 1.2-preview folder
instead. I both switched from > to >= as well as made the
comparison on the config/ folder instead of the options.xml
file to increase the chance of picking the really most
recently used folder.

Change-Id: If6d41c21de96d85b1a19e145701ba4f550938225