Make using Android Typed Integer renderer user settable
This CL provides a user option to enable using the Android Typed
Integer renderer for primitives. This can be enabled either from
"settings | debugger | Java", or by just right clicking on a
variable in the "Variables Panel" while debugging, and selecting
"Customize Data Views".

This change doesn't have to be merged back with JB. Without this, all
that happens is that the default renderers have precedence.

Change-Id: I906fc89073e4e1ac6afb00b1ddb0d288b4a37352