Ensure that the send usage statistics opt-in dialog works correctly
If the user has not yet answered whether or not they agree to sending
statistics, no statistics are sent - and the dialog is shown at
startup until the user decides one way or the other.

However, there was a problem with this code: it often resulted in the
dialog not being shown.

The root cause is that when the SendStatisticsComponent component runs
at startup, and it finds that it needs to ask the user, it creates the
notification and puts it into the notification bus.

However, it turns out that the notification manager, which normally
subscribes to notification events, can sometimes be initialized
*after* the statistics component (because they are both application
components, initialized by the IDE infrastructure from separate plugin
registration points, and the order appears to sometimes vary.)

This means that the event on the notification bus is never noticed and
no dialog appears. The "send statistics" state remains "ask user", no
statistics are sent, and on the next startup, the above bug can repeat
itself over and over.

The fix is simple: in the statistics application component, force
initialization of the notification manager application component
first.

Change-Id: I3f07b0966989a0deefbec7c8379ca9a73e953127
(cherry picked from commit 018cb311c73706671b3d61ab226eb45ca6433794)