Properly select the vmoptions file to edit when changing JVM flags.
Native launchers use system property jb.vmOptionsFile to pass the
(comma-separated) paths for all vmoptions files. Previously, the
VMOptions code assumed a single file was used. When an OutOfMemory
exception was detected, the user could tweak JVM flags such as Xmx but:
  i) it silently failed to persist the change if a custom vmoptions
     was already used;
 ii) it modified Studio's vmoptions.

This CL makes sure we always edit the custom, user-specified vmoptions,
which is always last and which always takes precedence over the default.

The code still has an implicit assumption that -Xmx, -XX:MaxPermSize and
-XX:ReservedCodeCacheSize are all present in the same vmoptions file.
It's possible the user adds just -Xmx1024 to the custom vmoptions, and
in this case the OutOfMemory warning dialog shows "unknown" for the other
fields.

Change-Id: I8dd06c5ff4cba4188a45526ed533464c16087ed5