Scroll adjustments can disengage/reengage Scroll To End
Using the mouse wheel or grabbing and moving the scrollbar in any
console now disengages/reengages scroll to end in the same way
moving the caret manually via keyboard would.

Related: https://youtrack.jetbrains.com/issue/IDEA-1R43939

This patch was checked into master at
https://github.com/JetBrains/intellij-community/commit/c8cec2f7243d0acfbdc259a62906efa49977a584

and we are submitting it here so it goes into our local copy of the
branch (but it should merge smoothly).

Change-Id: I5128976ce54fb06158781fce8ab6208be408bfb3