Use FullVersion for FullApplicationName.
FullApplicationName is used in places like the About box and the
IdeaLogger to indicate the running application. FullVersion is a string
like "Android Studio 1.4.0 Preview 3". This string is always preferable
to the build number ("AI-141.1234567") which is opaque.

Both About box and IdeaLogger include the build number even after using
the FullApplicationName.

Change-Id: I94568e7b40099e37a97d28cc6400bca42ade03f9