Include more information when a crash happens.
This CL includes more information when an exception happens. In
particular, this includes the OS name & version, JVM version, the
locale and the update channel the user is on.

Some of this was earlier reported through the user agent string. The
format of that UA string was not well specified. Rather than using that,
we use custom dimensions for all the information now. There is a limit
of 200 custom dimensions for a GA property, but it doesn't look like
we are in danger of coming anywhere close to that.

This CL also wraps the entire reporting action inside a try-catch block
in order to prevent any crashes from the reporter itself.

Change-Id: I7cada394411c955d1409267c7d93bb32a0fd22b7