Report summary quality metrics
This CL makes two changes to the exception reporter:

1. Whenever an exception happens, that fact is recorded in a file by
incrementing a counter. This is performed before the exception is
reported directly to GA in order to avoid further destabilizing the JVM.
This file is read back at start-up, and the counter is initialized
with that value. Such a scheme could potentially detect errors in the
exception reporter itself.

2. The count of exceptions and activity is sent out at periodic
intervals (once a minute after start-up, every 30 minutes later) to
tools.google.com. At a later point in time, we will remove the duplicate
POST about the activity count to GA.

Change-Id: Icb9f827a7cdbe4849f9e04119fcdbd072fc685a6