[UITests] Also use IDE classloader if running with a release build.
UI tests running with Ant on top of the source checkout are using a
single classloader, just as regular Intellij unit tests. Recreating
the actual classloader hierarchy is currently impossible due to
(explicit) dependencies on the Android plugin, such that the UI test
runner can't be loaded by anything else except the Android plugin's
classloader.

This is a temporary/hacky workaround recreating the single
classloader setup of unit tests until we refactor the UI test
framework.

Change-Id: I42a5730f55538440f46ebc19ff993aa29fb40620