Don't miss out on reporting total # of crashes
The count of fatal exceptions isn't persisted on disk. So if the
count is not reported before the IDE shuts down, then we miss out on
that data.

This CL changes the reporting to be done right when the file tracking
the crash itself is deleted.

Change-Id: Ia939258f4079df65fae9e07ade97a5e503c31098
(cherry picked from commit 31c9c9f74b5ba4795a1b7f32d8177d759fd05ee5)