[UITests] Bake the main class name into launchers, at build time.
Currently, we launch UI testing by overriding the main class in the
launchers with STUDIO_MAIN_CLASS_NAME. This works from the command-line
on Linux and Windows. On Mac it's not that simple: we need to change
Info.plist and then codesign. Also, IJ15 removes support for this
MAIN_CLASS_NAME environment variable.

With this CL, we set the main class at build time and no longer need to
depend on an extra environment variable when launching UI tests.

Change-Id: I319bd3d2d63c051574792afd061a48c2b5fd5b29