Synchronize reading/writing exception counts to files
We see exception counts reported every now and then that seem
quite high, and don't match up with other observed reports (GA).
This CL potentially fixes an issue where multiple bugs could
occur in different threads, and the persistence logic could
in theory mix them both up together. So we make the reading and
writing of the exception count file a critical section that
needs to be performed with a lock held.

Also adds a couple of assertions in internal mode to catch
possible errors.

Change-Id: I4f2009862238f4082b873d3a4cba95829c538804