184409: Disable recursive VFS refresh from FrameworkDetector
The VFS refresh code has a FrequentEventDetector setup such that
more than 100 refreshes within 100 ms triggers an error. We see
a number of feedback reports that indicate that this was triggered
from the FrameworkDetectionProcessor.

It is not clear how this code got triggered: It doesn't seem to
be invoked from any of our code. Regardless, this CL disables the
refresh call. This should be fine because a) we don't use this,
and b) the framework detection code is only to make things slightly
easier and doesn't affect real work.

Change-Id: I8889e083f7b491e1f09dd470a4e659765da14635
(cherry picked from commit 8c3a4fa4bb4a6d9162a8e6f90e05a99cbdb2419e)