Don't ask about dir vs package if only one non-generated source
If you have a single package source, which generates other folders in the same
package that are generated, the IDE will ask you whether you want to rename
the package or the directory.

If all but one of these folders are generated, don't ask - this is common
in Android.

Related: https://youtrack.jetbrains.com/issue/IDEA-125147
Motivating bug in Android Studio: http://b.android.com/184881

Change-Id: I8a221154fb27c7f2a2ca3f149bbda43d0d77f905