119540: Reduce the refresh rate of the threads panel
When you attempt to debug an app while the threads panel is open,
the application is simply stuck at the "Waiting for debugger.."
dialog.

This happens because android.os.Debug#waitForDebugger() waits for
about 1.3 seconds of no activity from the debugger once it has
connected, with the understanding that if the debugger is idle, then it
has finished its initialization (setting breakpoints, etc).

The threads panel however issues a refresh request every 200 ms. As a
result, the VM gets stuck in the waitForDebugger() routine since there
is never a 1.3 seconds of idle time.

This CL does a simple fix: we reduce the refresh rate of the threads
pane to once every 1.5 seconds. That seems just about as arbitrary as
a refresh every 200 ms..

Change-Id: Ida3335180333e8694c3516fb839936c45fb2009a