Move analytics upload code back into the platform.
In a previous CL, we moved all the code to upload stats out of
the platform into the analytics module. This was done in order to
share code between the platform and the android plugin.

This however resulted in a regression: Any exceptions/crashes
that need to be reported before the plugins are loaded would fail
because the extension point had not been loaded yet.

As a result, this CL moves back all the GA upload code back into
the platform. We still have the analytics module, which now exists
simply to keep the android plugin not dependent on the new APIs
present in the platform.

Change-Id: I9b3adbae60e93d5f0c16f36cb14c9a0bb51ea5a7
(cherry picked from commit e182bf2298c00b5a9b6cc005acfd988127412516)