162448: Don't copy certain awt properties from Studio to Gradle daemon.
The root cause is a Gradle bug which results in unnecessarily copying
system properties from Studio's JVM to Gradle daemon. Some of these
properties won't work in the daemon JVM, such as the AWT ones denoting
implementation classes (because they differ between Apple and
OpenJDK/Oracle JDKs).

This is meant as a patch until the Gradle side is fixed.

Change-Id: Ic8298f94413b5bf820cf77d48a2da28c9f3d97e3