Crash reporter: update record file with version string.
When we create the file, we don't yet have the version number, so
we write the build number instead (because a small fraction of
corrupted installations etc. will fail to go much further than
that).

We now replace the build number with the friendlier version string
when Studio gets past the bootstrap phase.

Change-Id: I98a1b1226542c6c675d31d4364bd2bc15f60edae