Exclude Log.error without an exception from exception counts
For exception tracking, we only want to count real exceptions.
We want to exclude code that does Log.error(msg), where it is just
trying to report an error (usually recoverable error conditions).

This CL achieves that by making Log.error use a specific throwable
instead of java.lang.Throwable, and not counting that specific
throwable.

Bug: 24000263
Change-Id: I4f6edc58097f82f41cbca147fe319d7558551821