[hidpi] Use alternate dpi detection mechanism on Linux
On Linux, the preferred/modern way of scaling applications according to
DPI settings seems to look at the Gtk font DPI property. Swing/AWT already
use this information to create PangoFonts. We still fallback to system
DPI when the desktop property is not found.

This was tested on Ubunty 14.04, by changing the "Text Scaling factor" in
the "System Setttings > Fonts" applet.

Change-Id: I7a41fa68a4ff0255de377c4235f6c51d282a273b