[hidpi] Add support for overriding system dpi with system property
The "hidpi.system.dpi.override" system property can now be used to
force a custom scaling factor. For example, by setting a value of
"144", a scaling factor of 1.5 is applied to the size of UI elements.

Note that this setting does not affect the scaling of the pre-defined
Swing fonts.

Change-Id: I94702d9955b0022e142065d02137f67c45447da6