Remove javaFX from the distribution
We had already explicitly removed the javaFX plugin (among many
others) in the excludedPlugins property in studio_properties.gant,
but at some point via an IJ merge we received a new custom plugin copy
section in layouts.gant which inserted it anyway. Remove it now
since this is unrelated to Android development.

Change-Id: Iec69736a778ae4d8e5363fe8940ee9c5fcfdd391