IDEA 12-14 to IDEA 15 Merge Reset
This changeset transfers a large number of individual CLs that didn't
merge cleanly from one branch to another.

SOURCE CODE ARCHAEOLOGY
-----------------------

To find the original CL and commit message for a given file and line,
follow these steps

$ git blame <file>      # => points to this CL
$ git checkout studio-1.4-release
$ git blame <same file>

The repository should now be in the pre-merge state and you
can explore the history.

When done, git checkout your previous branch.

WHY DID WE DO THIS?
-----------------------

When we started merging in IntelliJ CLs, we were flattening all CLs in
the git repository into a single CL for each EAP tag we were merging.
That meant that our IDE history only pointed to merge CLs, it didn't
have any context from the individual CLs that had been merged in the
IDE.

This was unfortunate, and made merging harder as well. To fix this,
for the IntelliJ 15 merge, we decided to completely rewrite our git
history to match IntelliJ's git history CL for CL instead, and then to
move our own modifictions on top of that.

Roughly 100 CLs were able to be cherrypicked directly, but many other
CLs, including anything that touched the build scripts folder, or
overlap areas where we'd upstreamed changes (such as in the updater)
or otherwise frequently updated (such as the system health monitor),
those weren't easiy to merge, so instead we just flatten them all into
this single CL.

Going forward we'll have clean histories both from our own changes,
and from all the IDE changes. For those cases where you need to chase
down the reason for a specific diff and it points to this CL, just
follow the above instructions for how to cross the merge boundary.

Change-Id: I73284e06a81aa25900a179587ddfc6c3dd3b80e8