188195: Add -XX:-OmitStackTraceInFastThrow to vmoptions
This is an attempt to track down a commonly occuring AIOOBE that
doesn't contain a stacktrace. Oracle release notes say:

"The compiler in the server VM now provides correct stack backtraces for
all "cold" built-in exceptions. For performance purposes, when such an
exception is thrown a few times, the method may be recompiled. After
recompilation, the compiler may choose a faster tactic using
preallocated exceptions that do not provide a stack trace. To disable
completely the use of preallocated exceptions, use this new flag:
-XX:-OmitStackTraceInFastThrow".

Change-Id: Idaa07281dda9a7f5dcd01a7dfe91d1a00bbb049b