Fix Accessible implementation of SimpleColoredComponent
SimpleColoredComponent is more of a LABEL than a simple generic
AWT_COMPONENT. Also, by inheriting from AccessibleJComponent, the
component expose default implementation for more methods. Finally,
by returning a "getAccessibleName", we offer a text to text readers.

This change was initially aimed at enabling screen readers in the
"ProjectView" tree, but it ended up applying in a lot more places,
as SimpleColoredComponent is very commonly used.

Change-Id: I960177fb2b4966aaf5183e93130b833650c93094