Restore check for write action being executed in EDT.
The check is present verbatim in ij15, and it looks like on our side
has been lost in commit 528b35b5 (back then it was
"assertCanRunWriteAction(getStatus())").

I noticed this while debugging UI tests and naively calling a write
action from the main thread: there were no warnings, it proceeded
with acquiring the lock and then blocked waiting for something that
never happened.

Change-Id: Ice858c357c6a430efa6a79d40f0b208b05e25e80