Track counts of Throwables from bundled and non-bundled plugins
Which plugin a Throwable is from, if any, is determined by
IdeErrorsDialog.findPluginId, which returns the result of
PluginManagerCore.getPluginByClassName for the first element of the
Throwable's stack trace for which PluginManagerCore.isPluginClass
is true, or null otherwise. (There are some other special cases;
see the code.)

Whether a plugin is “bundled” is determined by
IdeaPluginDescriptorImpl.isBundled, which returns true if it's under
Studio's “home” directory, or false otherwise. Non-bundled plugins
are normally under the config directory.

Change-Id: Ic5bc41f21ba049b4fa2280858b99c914d079d3d9