stop action: Don't include cancellable processes into consideration
In IntelliJ, the stop icon explicitly takes the running configurations
and the active background tasks into consideration while attempting to
figure out whether it should be enabled or not.

This change removes the cancellable background tasks from consideration
in this matter as we don't want things like Gradle compilations to have
any effect on the stop button.

Change-Id: I5adf0eb76cdbb88baeae008a944fb288905676e3