StopAction: Prefer the result from KillableProcess interface
The canBeStopped method in StopAction says that a process can be stopped
if either:
 - it hasn't terminated already
 - or if it implements the KillableProcess interface, then it looks at
   whether it actually says the process can be killed.

Because of the way the logic is structured, even if a process is not
killable, this method returns that it can be stopped. This CL fixes it
so that the result of the killable interface takes precedence.

Change-Id: I83c5325ee194e20c54562faa477f38721fb99dcf