The last opened project location was stored and loaded incorrectly.
Fixes issue: https://code.google.com/p/android/issues/detail?id=171375

The last opened project location was stored on the application properties
but not on the project properties the first time.
This fix makes sure that if the last project location was not loaded
in the project properties it will try and get it from the application
properties and save it to the project properties before going on.

Change-Id: Icea74976bda94c7bd60d6db436497dc58dbe3b67