Use a JDialog with a null owner for floating tool windows on Mac only.
On Windows and Ubuntu a missing owner on JDialog (or using a JFrame)
has the sad side effect of creating multiple icons in the system tool
bar, which makes is hard to bring all Android Studio windows to the
top (you have to do this 1 by 1).

On Mac however this fixes the issue of the tool window jumping from
one monitor to another, and on Mac there is only 1 icon in the dock
which brings all the tool windows and the main window to the top
with 1 click.

Bug: http://b.android.com/193168

Change-Id: Ie8e99756594902ab1d7ded03324775c769f4039c