a11y: Fix bad ij15 merge: nodes needs to have the "label" role
It seems there was a merge conflict wrt to the
implementation of Accessible on SimpleColoredComponent. In addition
to returning the text of the component, we also need to return
the "label" role. This is required by nvda do correctly detect
events related to node selection change (e.g. in Project View).

Change-Id: Ib098af9f8fd46c18fb62d7349ceb621ddc7d3a46