Experiment: remove CidrCommon prefix to unblock our unit tests.
Recent upstream changes have improved platform autodetection in the
test framework. Unfortunately, our setup includes the CidrCommon
module, so the autodetection mechanism will always set the platform
prefix to "CidrCommon" instead of "Idea" (even though we try to set
it via a JVM flag in the build script!).

It's unclear whether this is the right fix in the long term, mostly
because we don't fully understand the goals of platform autodetection.

Change-Id: I8d99f0daa2e12a1634175cb8c4512264b52d56ac