Add ability to override the m2 repository packaged in command-line build
We want this so that people can embed a different repo in the .zip/.tar.
This is used in two scenarios:
-not having to remember to update the maven repository in every project
-making it easier to build a self-contained version of Android Studio
 with updated gradle plugin for distribution to other machines (e.g OS
 or sharing with other people)

By setting the STUDIO_CUSTOM_REPO environment variable, the specified m2 repo
is used instead of the default  'google/vendor/dogfood/offline-m2'.
Such repo can be provisioned using './gradlew zipOffineRepo'

Change-Id: Ie24e25cb1de182de1f73efac2d33c6ef437d962b