a11y: Avoid infinite loop in getAccessibleParent.
Because we wrap accessible components with a
ExpandedItemRendererComponentWrapper, we end up having a cycle
where wrapper.getAccessibleContext().getAccessibleParent() == wrapper.

The reason is that wrapper.getAccessibleContext() returns
wrapper.child.getAccessibleContext().

This cycle is a problem because the Access Bridge sometimes uses
getAccessibleParent() in loops to find the root component. By having
a cycle, we make the Access Bridge run into infinite loops.

The workaround is to set the accessible parent of the child to
be the parent of the wrapper (which is sometimes null, that's ok).

Change-Id: I488668cfd70e17baee65d423911992f6e0997dc9