Fix Maximize window button when using JDK1.6 or JDK1.7
Bug: http://b.android.com/199298

This problem was caused by a bugfix to a different resize bug:
  https://youtrack.jetbrains.com/issue/IDEA-94602
The problem with the fix is that Window.setBounds is called
which will remove the maximize state just set when the user
maximizes Android Studio.
This fix will avoid applying the above fix when the window is
maximized.

Change-Id: Iafe23626721eb27ab6f2b30e9d06cd63a9ead840
(cherry picked from commit 5e75b245303923d84022164149e0bd81ce86c7b3)