Remove Restart Activity shortcut override
We added this back when restarting the activity was not
the default for a hotswap operation, and we thought this
needed to be done frequently. Now that restarting activities
on hot swap is the default, remove the shortcut and hand
it back to replace-in-path.

This reverts commit c5aa5488e5c740691bbec1fcd9eafa98da69181b.
(cherry picked from commit 6fec511f4ae7ed186ce5b82efb8af0cb07a24033)

Change-Id: Ib96373203cb915017e7bfac31441f0e20483033a